chore: Update dependencies
refactor: Utils#sleepSeconds api break change

+++ stun4j-guid-core/src/main/java/com/stun4j/guid/core/utils/Utils.java
@@ -38,7 +38,7 @@ public final class Utils {
     LockSupport.parkNanos(ms * NANOS_PER_MS);
   }
 
-  public static void sleepSeconds(long seconds) {
+  public static void sleepSeconds(int seconds) {
     LockSupport.parkNanos(seconds * NANOS_PER_SECONDS);
   }
